# design/openapi.yaml
openapi: 3.0.3
info:
  title: EventPulse API
  description: |
    A comprehensive event management and analytics platform built with Spring Boot. 
    EventPulse provides real-time event ingestion, processing, and analytics capabilities 
    with JWT authentication, WebSocket support, and detailed metrics.
  version: 1.0.0
  contact:
    name: EventPulse Team
    email: support@eventpulse.com
    url: https://eventpulse.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.eventpulse.com
    description: Production Server
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User Authentication
      description: Authenticate a user with username and password, returns a JWT token for accessing protected endpoints.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Admin login
                value:
                  username: admin
                  password: password
              user:
                summary: Regular user login
                value:
                  username: user
                  password: password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/validate:
    get:
      tags:
        - Authentication
      summary: Validate JWT Token
      description: Validate an existing JWT token and return token information if valid.
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzUxMiJ9..."
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login endpoint
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Username for authentication
          example: "admin"
        password:
          type: string
          minLength: 6
          maxLength: 100
          description: Password for authentication
          example: "password"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
          example: "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImlhdCI6MTY3MDAwMDAwMCwiZXhwIjoxNjcwMDg2NDAwfQ.example-signature"
        type:
          type: string
          description: Token type
          example: "Bearer"
        username:
          type: string
          description: Authenticated username
          example: "admin"
        expiresAt:
          type: string
          format: date-time
          description: Token expiration time
          example: "2023-12-02T10:30:00Z"
        message:
          type: string
          description: Response message
          example: "Authentication successful"
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2023-12-01T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid request data"
        details:
          type: object
          description: Additional error details
          additionalProperties:
            type: string
          example:
            source: "Event source is required"
            type: "Event type must not exceed 100 characters"